Committing the latest fixes which improves the accuracy of reordering and event duplication

+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/query/processor/window/KslackWindowProcessor.java
@@ -95,7 +95,7 @@ public class KslackWindowProcessor extends WindowProcessor {
 
         List<InEvent> newEventList = new ArrayList<InEvent>();
         try {
-            long timestamp = (Long) event.getTimeStamp();
+            long timestamp = event.getTimeStamp();
             if (expireFlag) {
                 if (timestamp < lastSentTimeStamp) {
                     return;
